♻️ refactor : investments에 있는 header 데이터를 따로뺌
