+++ OneCoin/Client/src/utills/constants/header.ts
@@ -1,0 +1,13 @@
+export const HEADER_LIST = {
+	page: [
+		{ name: '거래소', path: '/exchange' },
+		{ name: '입금', path: '/balances' },
+		{ name: '투자내역', path: '/investments/balance' },
+		{ name: '스왑', path: '/swap' },
+	],
+	logout: [
+		{ name: '로그인', path: '/login' },
+		{ name: '회원가입', path: '/signup' },
+	],
+	login: [{ name: '마이페이지', path: '/mypage' }],
+};
